#585 Move the system exit catcher security manager closer to its usage package (state).
